Change BufferHelper::reqMemSizePercent to double for higher precision

+++ common/include/lima/BufferHelper.h
@@ -55,7 +55,7 @@ class LIMACORE_API BufferHelper
 		bool initMem;
 		DurationPolicy durationPolicy;
 		PersistentSizePolicy sizePolicy;
-		int reqMemSizePercent;
+		double reqMemSizePercent;
 
 		Parameters();
